move svg files, update paths
